// packages/sdk/src/zod/generated/AdminImage.ts
// This file is auto-generated. Do not edit directly.
import { z } from 'zod';

export const AdminImageSchema = z.object({
  id: z.string(),
  viewable_id: z.string(),
  type: z.string().nullable(),
  viewable_type: z.string(),
  position: z.number().nullable(),
  alt: z.string().nullable(),
  created_at: z.string(),
  updated_at: z.string(),
  original_url: z.string().nullable(),
  mini_url: z.string().nullable(),
  small_url: z.string().nullable(),
  medium_url: z.string().nullable(),
  large_url: z.string().nullable(),
  xlarge_url: z.string().nullable(),
  og_image_url: z.string().nullable(),
});

export type AdminImage = z.infer<typeof AdminImageSchema>;